v.1.15.0.2

+++ engine/src/wolf.render/w_graphics/w_gui_render.cpp
@@ -179,7 +179,7 @@ namespace wolf
                                 _count * sizeof(wolf::gui::w_gui_vertex_2d),
                                 _count,
                                 _indices.data(),
-                                _indices.size()) == S_FALSE)
+                                static_cast<UINT>(_indices.size())) == S_FALSE)
                             {
                                 _hr = S_FALSE;
                                 V(_hr, "loading mesh", this->_name, 3);
